Revert "Revert "Revert "Enhance UI with new layout components, menu functionality, and improv…"""
