+++ Stellar-View/src/App.tsx
@@ -1,10 +1,10 @@
 //import React from 'react';
 import { BrowserRouter as Router, Routes, Route, } from "react-router-dom";
-import Landing from "./Pages/Landing.tsx"
+import Landing from "./Pages/Landing/Landing.tsx";
 import './App.css'
-import ViewPage from "./Pages/ViewPage.tsx";
-import ImageView from "./Pages/ImageView.tsx";
-import ExperimentalImageView from "./Pages/ExperimentalView.tsx";
+import ViewPage from "./Pages/View/ViewPage.tsx";
+import ImageView from "./Pages/Imageview/ImageView.tsx";
+
 function App() {
 
     return (
@@ -13,7 +13,6 @@ function App() {
                 <Route path="/" element={<Landing/>} />
                 <Route path="/planet-view" element={<ViewPage/>} />
                 <Route path="/image-view" element={<ImageView/>} />
-                <Route path="/experimental-view" element={<ExperimentalImageView/>} />
             </Routes>
         </Router>
     )
