+++ Stellar-View/src/Components/Layout.tsx
@@ -1,32 +1,24 @@
 import MaterialIcon from "./MaterialIcon.tsx";
-import Menu from "./Menu.tsx";
-import {useState} from "react";
+
 interface LayoutProps {
-    children?: React.ReactNode,
-    title?: string
+    children?: React.ReactNode
+    //title?: string come back to possibly later
 }
 
-export default function Layout ({ children, title }: LayoutProps) {
-    const [menuOpen, setMenuOpen] = useState(false);
-
-    const HandleMenuState = () => {
-        setMenuOpen(!menuOpen);
-    }
-
+export default function Layout ({ children }: LayoutProps) {
     return(
         <>
             <div
                 className="min-w-screen bg-black flex p-2 items-center justify-end">
                 <div id="menu-container">
-                    {title && <span className="text-white mr-4 font-semibold">{title}</span>}
-                <span onClick={HandleMenuState} className="cursor-pointer p-5 ">
+                <span className="cursor-pointer hover:bg-purple-700 rounded-3xl transition-colors px-2 ">
+                    <MaterialIcon name="question_mark" color="white" />
+                </span>
+                <span className="cursor-pointer p-5 hover:bg-purple-700 rounded-3xl transition-colors ">
                     <MaterialIcon name="menu" color="white" />
                 </span>
                 </div>
             </div>
-            {menuOpen && (
-                <Menu SetMenuState={setMenuOpen} menu={menuOpen} />
-            )}
             <div id="child-content">
                 {children}
             </div>
