+++ Stellar-View/src/Pages/View/ViewPage.tsx
@@ -8,20 +8,6 @@ import {
     Color
 } from "cesium";
 import * as Cesium from "cesium";
-import Layout from "../Components/Layout.tsx"
-
-
-export type GIBSLayerConfig = {
-    url: string;
-    matrix: string;
-    layer: string;
-    name: string;
-    format: string;
-    maxLevel: number;
-    description: string;
-    temporal: "10min" | "daily" | "none";
-};
-
 const GIBS_LAYERS = {
     // GOES - 10 MINUTE updates! (Geostationary)
     goes_east_geocolor: {
@@ -131,14 +117,14 @@ const GIBS_LAYERS = {
         temporal: "daily",
     },
 };
-export default function ViewPage() { //fix typings
+export default function ViewPage() {
     const cesiumContainer = useRef<HTMLDivElement | null>(null);
     const [viewer, setViewer] = useState<Viewer | null>(null);
-    const [activeLayer, setActiveLayer] = useState<Cesium.ImageryLayer | null>(null);
+    const [activeLayer, setActiveLayer] = useState<any>(null);
     const [currentDate, setCurrentDate] = useState(() => {
         // Default to yesterday (data usually available with 1 day delay)
         const yesterday = new Date();
-        yesterday.setDate(yesterday.getDate() - 3);
+        yesterday.setDate(yesterday.getDate() - 2);
         return yesterday.toISOString().split('T')[0];
     });
     const [currentTime, setCurrentTime] = useState("12:00");
@@ -146,6 +132,9 @@ export default function ViewPage() { //fix typings
     const [resolution, setResolution] = useState<string>("medium");
 
     const [selectedPlanet, setSelectedPlanet] = useState<string>("earth");
+
+
+// Add this ref to track cleanup
     const viewerRef = useRef<Viewer | null>(null);
 
     useEffect(() => {
@@ -165,128 +154,126 @@ export default function ViewPage() { //fix typings
             (async () => {
                 const provider = await Cesium.createWorldImageryAsync();
                 v.imageryLayers.addImageryProvider(provider);
-
-                viewerRef.current = v;
-                v.camera.setView({
-                    destination: Cartesian3.fromDegrees(0, 0, 20000000), // Zoomed out to see the whole Earth
-                    orientation: {
-                        heading: 0,
-                        pitch: -Math.PI / 2,
-                        roll: 0
-                    }
-                });
-
-                v.scene.globe.backFaceCulling = true;
-                v.entities.add({
-                    name: "GOES-West (GOES-17)",
-                    position: Cesium.Cartesian3.fromDegrees(
-                        -137.2, // longitude (W = negative)
-                        0,      // latitude (over equator)
-                        35786000 // altitude in meters (~35,786 km)
-                    ),
-                    point: {
-                        pixelSize: 12,
-                        color: Cesium.Color.ORANGE,
-                        outlineColor: Cesium.Color.WHITE,
-                        outlineWidth: 2,
-                    },
-                    label: {
-                        text: "GOES-West",
-                        font: "14px sans-serif",
-                        fillColor: Cesium.Color.WHITE,
-                        pixelOffset: new Cesium.Cartesian2(0, -20),
-                    }
-                })
-                v.entities.add({
-                    name: "GOES-East (GOES-16)",
-                    position: Cesium.Cartesian3.fromDegrees(
-                        -75.2, // longitude (W = negative)
-                        0,     // latitude (over equator)
-                        35786000 // altitude in meters (~35,786 km)
-                    ),
-                    point: {
-                        pixelSize: 12,
-                        color: Cesium.Color.CYAN,
-                        outlineColor: Cesium.Color.WHITE,
-                        outlineWidth: 2,
-                    },
-                    label: {
-                        text: "GOES-East",
-                        font: "14px sans-serif",
-                        fillColor: Cesium.Color.WHITE,
-                        pixelOffset: new Cesium.Cartesian2(0, -20),
-                    }
-                });
-
-                // Add animated satellites
-                const createModisSatellite = (name: string, color: Cesium.Color, phaseOffset: number) => {
-                    return v.entities.add({
-                        name: name,
-                        position: new Cesium.CallbackProperty(() => {
-                            const time = v.clock.currentTime;
-                            const date = Cesium.JulianDate.toDate(time);
-
-                            // Orbital period: 98.8 minutes
-                            const orbitalPeriodMinutes = 98.8;
-                            const minutesSinceEpoch = (date.getTime() / 60000) + phaseOffset;
-                            const orbitalAngle = (minutesSinceEpoch / orbitalPeriodMinutes) * 2 * Math.PI;
-
-                            // Earth rotation - satellites drift westward due to sun-synchronous orbit
-                            const earthRotationRate = (2 * Math.PI) / (24 * 60); // radians per minute
-                            const earthRotation = minutesSinceEpoch * earthRotationRate;
-
-                            const altitude = 705000;
-                            const earthRadius = 6371000;
-                            const inclination = Cesium.Math.toRadians(98.2);
-                            const radius = earthRadius + altitude;
-
-                            // Position in orbital plane
-                            const xOrbit = radius * Math.cos(orbitalAngle);
-                            const yOrbit = radius * Math.sin(orbitalAngle);
-
-                            // Rotate for inclination
-                            const x = xOrbit;
-                            const y = yOrbit * Math.cos(inclination);
-                            const z = yOrbit * Math.sin(inclination);
-
-                            // Apply Earth rotation (rotate around Z-axis)
-                            const xFinal = x * Math.cos(-earthRotation) - y * Math.sin(-earthRotation);
-                            const yFinal = x * Math.sin(-earthRotation) + y * Math.cos(-earthRotation);
-                            const zFinal = z;
-
-                            return new Cesium.Cartesian3(xFinal, yFinal, zFinal);
-                        }, false) as unknown as Cesium.PositionProperty,
-                        point: {
-                            pixelSize: 10,
-                            color: color,
-                            outlineColor: Cesium.Color.WHITE,
-                            outlineWidth: 2,
-                        },
-                        label: {
-                            text: `${name}\n(EOS ${name === "MODIS Terra" ? "AM-1" : "PM-1"})`,
-                            font: "12px sans-serif",
-                            fillColor: Cesium.Color.WHITE,
-                            pixelOffset: new Cesium.Cartesian2(0, -20),
-                            showBackground: true,
-                            backgroundColor: Cesium.Color.BLACK.withAlpha(0.7),
-                        }
-                    });
-                };
-
-                createModisSatellite("MODIS Terra", Cesium.Color.CYAN, 0);
-
-                createModisSatellite("MODIS Aqua", Cesium.Color.LIGHTBLUE, 180);
-                // Configure clock f        te range
-                v.clock.startTime = JulianDate.fromIso8601("2000-01-01T00:00:00Z");
-                v.clock.stopTime = JulianDate.fromIso8601(new Date().toISOString());
-                v.clock.currentTime = JulianDate.fromIso8601(currentDate + "T" + currentTime + ":00Z");
-                v.clock.clockRange = 0;
-                v.clock.multiplier = 600;
-
-
-                setViewer(v);
             })();
             viewerRef.current = v;
+            v.camera.setView({
+                destination: Cartesian3.fromDegrees(0, 0, 20000000), // Zoomed out to see the whole Earth
+                orientation: {
+                    heading: 0,
+                    pitch: -Math.PI / 2,
+                    roll: 0
+                }
+            });
+
+            v.scene.globe.backFaceCulling = true;
+            v.entities.add({
+                name: "GOES-West (GOES-17)",
+                position: Cesium.Cartesian3.fromDegrees(
+                    -137.2, // longitude (W = negative)
+                    0,      // latitude (over equator)
+                    35786000 // altitude in meters (~35,786 km)
+                ),
+                point: {
+                    pixelSize: 12,
+                    color: Cesium.Color.ORANGE,
+                    outlineColor: Cesium.Color.WHITE,
+                    outlineWidth: 2,
+                },
+                label: {
+                    text: "GOES-West",
+                    font: "14px sans-serif",
+                    fillColor: Cesium.Color.WHITE,
+                    pixelOffset: new Cesium.Cartesian2(0, -20),
+                }
+            })
+            v.entities.add({
+                name: "GOES-East (GOES-16)",
+                position: Cesium.Cartesian3.fromDegrees(
+                    -75.2, // longitude (W = negative)
+                    0,     // latitude (over equator)
+                    35786000 // altitude in meters (~35,786 km)
+                ),
+                point: {
+                    pixelSize: 12,
+                    color: Cesium.Color.CYAN,
+                    outlineColor: Cesium.Color.WHITE,
+                    outlineWidth: 2,
+                },
+                label: {
+                    text: "GOES-East",
+                    font: "14px sans-serif",
+                    fillColor: Cesium.Color.WHITE,
+                    pixelOffset: new Cesium.Cartesian2(0, -20),
+                }
+            });
+
+                // Add animated satellites
+            const createModisSatellite = (name: string, color: Cesium.Color, phaseOffset: number) => {
+                return v.entities.add({
+                    name: name,
+                    position: new Cesium.CallbackProperty(() => {
+                        const time = v.clock.currentTime;
+                        const date = Cesium.JulianDate.toDate(time);
+
+                        // Orbital period: 98.8 minutes
+                        const orbitalPeriodMinutes = 98.8;
+                        const minutesSinceEpoch = (date.getTime() / 60000) + phaseOffset;
+                        const orbitalAngle = (minutesSinceEpoch / orbitalPeriodMinutes) * 2 * Math.PI;
+
+                        // Earth rotation - satellites drift westward due to sun-synchronous orbit
+                        const earthRotationRate = (2 * Math.PI) / (24 * 60); // radians per minute
+                        const earthRotation = minutesSinceEpoch * earthRotationRate;
+
+                        const altitude = 705000;
+                        const earthRadius = 6371000;
+                        const inclination = Cesium.Math.toRadians(98.2);
+                        const radius = earthRadius + altitude;
+
+                        // Position in orbital plane
+                        const xOrbit = radius * Math.cos(orbitalAngle);
+                        const yOrbit = radius * Math.sin(orbitalAngle);
+
+                        // Rotate for inclination
+                        const x = xOrbit;
+                        const y = yOrbit * Math.cos(inclination);
+                        const z = yOrbit * Math.sin(inclination);
+
+                        // Apply Earth rotation (rotate around Z-axis)
+                        const xFinal = x * Math.cos(-earthRotation) - y * Math.sin(-earthRotation);
+                        const yFinal = x * Math.sin(-earthRotation) + y * Math.cos(-earthRotation);
+                        const zFinal = z;
+
+                        return new Cesium.Cartesian3(xFinal, yFinal, zFinal);
+                    }, false) as unknown as Cesium.PositionProperty,
+                    point: {
+                        pixelSize: 10,
+                        color: color,
+                        outlineColor: Cesium.Color.WHITE,
+                        outlineWidth: 2,
+                    },
+                    label: {
+                        text: `${name}\n(EOS ${name === "MODIS Terra" ? "AM-1" : "PM-1"})`,
+                        font: "12px sans-serif",
+                        fillColor: Cesium.Color.WHITE,
+                        pixelOffset: new Cesium.Cartesian2(0, -20),
+                        showBackground: true,
+                        backgroundColor: Cesium.Color.BLACK.withAlpha(0.7),
+                    }
+                });
+            };
+
+            createModisSatellite("MODIS Terra", Cesium.Color.CYAN, 0);
+
+            createModisSatellite("MODIS Aqua", Cesium.Color.LIGHTBLUE, 180);
+            // Configure clock for better date range
+            v.clock.startTime = JulianDate.fromIso8601("2000-01-01T00:00:00Z");
+            v.clock.stopTime = JulianDate.fromIso8601(new Date().toISOString());
+            v.clock.currentTime = JulianDate.fromIso8601(currentDate + "T" + currentTime + ":00Z");
+            v.clock.clockRange = 0;
+            v.clock.multiplier = 600;
+
+
+            setViewer(v);
         } else if (selectedPlanet === "mars") {
             //3644333
             v = new Viewer(cesiumContainer.current, {
@@ -523,18 +510,18 @@ export default function ViewPage() { //fix typings
 
         const layerConfig = GIBS_LAYERS[selectedLayerKey as keyof typeof GIBS_LAYERS];
 
-        // TIME API FORMAT HAS TO BE EXACTLY RIGHT OR IT 404s IT MAKES SENSE BUT WHYYYYY even a millisecond off 404s and doesnt tell you what is wrong
-        // AT LEAST GIVE US A HINT
+        // For 10-minute data, format time as YYYY-MM-DDTHH:MM:SSZ
+        // For daily data, format as YYYY-MM-DD
         let timeString;
         if (layerConfig.temporal === "10min") {
             timeString = `${currentDate}T${currentTime}:00Z`;
-            //snap time to 10 minute increments to avoid missing api call on 10 minute imagry, so dumb
+            //snap time to 10 minute increments to avoid missing api call on 10 minute imagry
             const dateObj = new Date(`${currentDate}T${currentTime}:00Z`);
             const minutes = Math.floor(dateObj.getUTCMinutes() / 10) * 10;
             const rounded = new Date(dateObj);
             rounded.setUTCMinutes(minutes, 0, 0);
             timeString = rounded.toISOString().split(".")[0] + "Z";
-
+            // → e.g. "2025-09-30T12:10:00Z"
         } else {
             timeString = currentDate;
         }
@@ -572,12 +559,23 @@ export default function ViewPage() { //fix typings
             const dateString = date.toISOString().split('T')[0];
             //const timeString = date.toISOString().split('T')[1].substring(0, 5); // HH:MM
 
-            //const layerConfig = GIBS_LAYERS[selectedLayerKey as keyof typeof GIBS_LAYERS];
+            const layerConfig = GIBS_LAYERS[selectedLayerKey as keyof typeof GIBS_LAYERS];
 
+            if (layerConfig.temporal === "10min") {
+                // For 10-minute data, round to nearest 10 minutes
+                const minutes = Math.floor(date.getMinutes() / 10) * 10;
+                const roundedTime = `${String(date.getHours()).padStart(2, '0')}:${String(minutes).padStart(2, '0')}`;
+
+                if (dateString !== currentDate || roundedTime !== currentTime) {
+                    setCurrentDate(dateString);
+                    setCurrentTime(roundedTime);
+                }
+            } else {
+                // For daily data, only update if date changed
                 if (dateString !== currentDate) {
                     setCurrentDate(dateString);
                 }
-
+            }
         });
 
         return () => {
@@ -592,11 +590,25 @@ export default function ViewPage() { //fix typings
         };
 
     }, [viewer, currentDate, currentTime, selectedLayerKey]);
+/*
+    const handleDateChange = (e: React.ChangeEvent<HTMLInputElement>) => {
+        setCurrentDate(e.target.value);
+    };
+*/
+
     const handleLayerChange = (e: React.ChangeEvent<HTMLSelectElement>) => {
         setSelectedLayerKey(e.target.value);
     };
-
+/*
+    const toggleLayer = () => {
+        if (activeLayer) {
+            activeLayer.show = !activeLayer.show;
+            setActiveLayer({...activeLayer});
+        }
+    };
+*/
     const HandleResolutionChange = (e: React.ChangeEvent<HTMLSelectElement>) => {
+        //resChangedRef.current = true;
         setResolution(e.target.value);
     }
 
@@ -615,13 +627,13 @@ export default function ViewPage() { //fix typings
         setSelectedPlanet(planet);
     }
     return (
-        <Layout title="Daily Imagery: Try the Experiemental version for 10 minute updates here ->">
-            <div className="relative w-full h-full bg-black">
+        <>
+            <div className="relative w-full h-screen bg-black">
                 <div ref={cesiumContainer} className="w-full h-full" />
 
-                <div className="absolute top-0 left-4 bg-black/90 text-white p-2 sm:p-2 rounded-lg space-y-3 w-1/3 max-w-xs sm:max-w-sm shadow-xl border border-gray-700">
-                    <h3 className="font-bold text-base sm:text-xl text-blue-400">NASA GIBS Earth Viewer</h3>
-                    <p>Fullscreen Recommended for PC - Landscape Recommended for Mobile </p>
+                <div className="absolute top-4 left-4 bg-black/90 text-white p-4 rounded-lg space-y-3 max-w-sm shadow-xl border border-gray-700">
+                    <h3 className="font-bold text-xl text-blue-400">NASA GIBS Earth Viewer</h3>
+
                     <div className="space-y-3">
                         <label className="block text-sm font-medium mb-1">
                             Planet:
@@ -747,6 +759,332 @@ export default function ViewPage() { //fix typings
                 ) : null}
 
             </div>
-        </Layout>
+        </>
     );
 }
+
+/*
+export default function ViewPage() {
+    const cesiumContainer = useRef(null);
+    const [viewer, setViewer] = useState<Viewer | null>(null);
+    const [activeLayer, setActiveLayer] = useState(null);
+    const [selectedLayerKey, setSelectedLayerKey] = useState("goes_east_geocolor");
+    const [currentTimeString, setCurrentTimeString] = useState("");
+    const [preloadedCount, setPreloadedCount] = useState(0);
+
+    const lastLayerTimeRef = useRef("");
+    const lastUpdateTimeRef = useRef(Date.now() - 1000);
+    const preloadCacheRef = useRef(new Map());
+    const isJumpingRef = useRef(false);
+    const preloadIntervalRef = useRef(null);
+
+    const [resolution, setResolution] = useState<string>("medium");
+    const resChangedRef = useRef<boolean>(false);
+    const [isChangingLayer, setIsChangingLayer] = useState<boolean>(false);
+
+
+
+
+    useEffect(() => {
+        if (!viewer) return;
+        if (!activeLayer) return;
+
+        // Clear cache to force new providers with updated resolution
+        preloadCacheRef.current.clear();
+        console.log('cache ref', preloadCacheRef.current);
+        setPreloadedCount(0);
+        lastLayerTimeRef.current = "";
+
+        const layerConfig = GIBS_LAYERS[selectedLayerKey];
+        const date = JulianDate.toDate(viewer.clock.currentTime);
+        const timeString = getFormattedTime(date, layerConfig);
+
+        // Remove current layer
+        if (activeLayer) {
+            viewer.imageryLayers.remove(activeLayer);
+        }
+
+        // Create new provider with updated maxLevel
+        updateLayer(timeString);
+
+    }, [resolution]);
+    // Background preloader - loads past data continuously
+    useEffect(() => {
+        if (!viewer || isChangingLayer) return;
+
+        const layerConfig = GIBS_LAYERS[selectedLayerKey];
+
+        let preloadIndex = 1;
+        const maxPreload = layerConfig.temporal === "10min" ? 24 : 30;
+
+        const intervalId = setInterval(() => {
+            if (preloadIndex >= maxPreload) {
+                clearInterval(intervalId);
+                return;
+            }
+
+            const currentDate = JulianDate.toDate(viewer.clock.currentTime);
+            const preloadDate = new Date(currentDate);
+
+            if (layerConfig.temporal === "10min") {
+                preloadDate.setMinutes(preloadDate.getMinutes() - (preloadIndex * 10));
+            } else {
+                preloadDate.setDate(preloadDate.getDate() - preloadIndex);
+            }
+
+            const timeString = getFormattedTime(preloadDate, layerConfig);
+            preloadImagery(timeString, layerConfig);
+
+            setPreloadedCount(preloadIndex);
+            preloadIndex++;
+        }, 500);
+
+        preloadIntervalRef.current = intervalId;
+
+        return () => {
+            clearInterval(intervalId);
+            preloadIntervalRef.current = null;
+        };
+    }, [viewer, selectedLayerKey, isChangingLayer]);
+
+    // Listen to clock and update layer when time changes
+    useEffect(() => {
+        if (!viewer) return;
+
+        const layerConfig = GIBS_LAYERS[selectedLayerKey];
+
+        layerConfig.maxLevel = resolution === "potato" ? 1 : resolution === "low" ? 3 : resolution === "medium" ? 5 : layerConfig.maxLevel = 9;
+
+        const clockListener = viewer.clock.onTick.addEventListener((clock) => {
+            const now = Date.now();
+            const timeSinceLastUpdate = now - lastUpdateTimeRef.current;
+
+            // Detect if we're jumping (large time change in short period)
+            if (timeSinceLastUpdate < 100 && !isJumpingRef.current) {
+                isJumpingRef.current = true;
+                console.log("🚀 Jump detected - throttling updates");
+            } else if (timeSinceLastUpdate > 500) {
+                isJumpingRef.current = false;
+            }
+
+            const date = JulianDate.toDate(clock.currentTime);
+            const timeString = getFormattedTime(date, layerConfig);
+
+            setCurrentTimeString(timeString);
+
+            // If jumping, only update every 300ms to skip intermediate frames
+            if (isJumpingRef.current && timeSinceLastUpdate < 300) {
+                return;
+            }
+
+            // Only update layer if time actually changed
+            if (timeString !== lastLayerTimeRef.current && now - lastUpdateTimeRef.current > 500) {
+                updateLayer(timeString);
+                lastUpdateTimeRef.current = now;
+            }
+
+        });
+
+        return () => {
+            viewer.clock.onTick.removeEventListener(clockListener);
+        };
+    }, [viewer, selectedLayerKey, resolution]);
+
+
+    // Preload imagery provider into cache
+   const preloadImagery = (timeString: string, layerConfig) => {
+        const cacheKey = `${layerConfig.layer}-${timeString}-${resolution}`;
+
+        if (preloadCacheRef.current.has(cacheKey)) {
+            console.log("found cached", cacheKey)
+            return preloadCacheRef.current.get(cacheKey);
+        }
+
+        //update resolution max level
+        layerConfig.maxLevel = resolution === "potato" ? 1 : resolution === "low" ?  3 : resolution === "medium" ?  5 : layerConfig.maxLevel = 9;
+        // Create new imagery provider
+        const url = layerConfig.url.replace("{Time}", timeString);
+        const provider = new WebMapTileServiceImageryProvider({
+            url: url,
+            layer: layerConfig.layer,
+            style: "default",
+            tileMatrixSetID: layerConfig.matrix,
+            format: layerConfig.format,
+            credit: new Credit("NASA GIBS"),
+            maximumLevel: layerConfig.maxLevel,
+        });
+
+        preloadCacheRef.current.set(cacheKey, provider);
+        return provider;
+    };
+
+    // Update layer - use cached provider if available
+    const updateLayer = (timeString) => {
+        if (!viewer || isChangingLayer) return;
+
+        const layerConfig = GIBS_LAYERS[selectedLayerKey];
+        const cacheKey = `${layerConfig.layer}-${timeString}-${resolution}`;
+
+        //Remove existing layer
+        if (activeLayer && timeString !== lastLayerTimeRef.current) {
+            viewer.imageryLayers.remove(activeLayer);
+            console.log("✗ Removed previous imagery layer");
+        }
+
+        // Use cached provider or create new one
+        let provider;
+        console.log("Cache size:", preloadCacheRef.current.size);
+        console.log("Cache items:", preloadCacheRef.current);
+
+        if (preloadCacheRef.current.has(cacheKey)) {
+            provider = preloadCacheRef.current.get(cacheKey);
+            console.log("✓ Using cached imagery for:", timeString);
+       } else {
+            provider = preloadImagery(timeString, layerConfig);
+            console.log("⟳ Loading new imagery for:", timeString);
+       }
+
+        const layer = viewer.imageryLayers.addImageryProvider(provider);
+        layer.alpha = 0.8;
+        setActiveLayer(layer);
+        lastLayerTimeRef.current = timeString;
+    };
+
+
+    const handleLayerChange = (e) => {
+        setIsChangingLayer(true)
+
+        // Clear preload interval
+        if (preloadIntervalRef.current) {
+            clearInterval(preloadIntervalRef.current);
+            preloadIntervalRef.current = null;
+        }
+
+        // Clear cache and reset counters
+        preloadCacheRef.current.clear();
+        setPreloadedCount(0);
+        lastLayerTimeRef.current = "";
+
+        setSelectedLayerKey(e.target.value);
+
+        setTimeout(() => {
+            preloadCacheRef.current.clear();
+        setPreloadedCount(0);
+        setIsChangingLayer(false);
+        console.log('finished changing layer')
+        },7000);
+    };
+
+    const goToNow = () => {
+        if (viewer) {
+            const now = new Date();
+            viewer.clock.currentTime = JulianDate.fromDate(now);
+        }
+    };
+
+    const currentLayerInfo = GIBS_LAYERS[selectedLayerKey];
+    const isTemporalLayer = currentLayerInfo.temporal === "10min" || currentLayerInfo.temporal === "daily";
+
+    const HandleResolutionChange = (e: React.ChangeEvent<HTMLSelectElement>) => {
+        resChangedRef.current = true;
+        setResolution(e.target.value);
+        }
+    return (
+        <div className="relative w-full h-screen bg-black">
+            <div ref={cesiumContainer} className="w-full h-full" />
+
+            <div className="absolute top-4 left-4 bg-black/90 text-white p-4 rounded-lg space-y-3 max-w-sm shadow-xl border border-gray-700">
+                <h3 className="font-bold text-xl text-blue-400">NASA GIBS Earth Viewer</h3>
+
+                <div className="space-y-3">
+                    <div>
+                        <label className="block text-sm font-medium mb-1">
+                            Imagery Layer:
+                        </label>
+                        <select
+                            value={selectedLayerKey}
+                            onChange={handleLayerChange}
+                            className="w-full px-3 py-2 bg-gray-800 border border-gray-600 rounded text-white focus:border-blue-500 focus:outline-none"
+                        >
+                            <optgroup label="🔴 LIVE - 10 Minute Updates">
+                                <option value="goes_east_geocolor">GOES-East (Americas)</option>
+                                <option value="goes_west_geocolor">GOES-West (Pacific)</option>
+                            </optgroup>
+                            <optgroup label="True Color Imagery">
+                                <option value="modis_terra">MODIS Terra (Daily)</option>
+                                <option value="modis_aqua">MODIS Aqua (Daily)</option>
+                                <option value="viirs_snpp">VIIRS High Resolution (Daily)</option>
+                            </optgroup>
+                            <optgroup label="Snow & Ice">
+                                <option value="snow_cover">Snow Cover</option>
+                            </optgroup>
+                            <optgroup label="Temperature">
+                                <option value="land_temp_day">Land Temperature</option>
+                            </optgroup>
+                            <optgroup label="Fires & Hazards">
+                                <option value="fires">Active Fires</option>
+                            </optgroup>
+                            <optgroup label="Weather">
+                                <option value="cloud_fraction">Cloud Coverage</option>
+                            </optgroup>
+                            <optgroup label="Vegetation">
+                                <option value="ndvi">Vegetation Health (NDVI)</option>
+                            </optgroup>
+                        </select>
+                        <p className="text-xs text-gray-400 mt-1">
+                            {currentLayerInfo.description}
+                        </p>
+                    </div>
+
+                    <div>
+                        <button
+                            onClick={goToNow}
+                            className="w-full px-4 py-2 bg-blue-600 hover:bg-blue-700 rounded transition-colors"
+                        >
+                            Jump to Now
+                        </button>
+                    </div>
+                </div>
+
+
+                <div className="pt-2 border-t border-gray-700 text-xs space-y-1">
+                    <p className="text-gray-400">Data: NASA GIBS</p>
+                    {isTemporalLayer ? (
+                        <p className="text-green-400 font-medium">⚡ Live: Updates every 10 minutes</p>
+                    ) : (
+                        <p className="text-gray-400">Updated within 3 hours of observation</p>
+                    )}
+                    {currentTimeString && (
+                        <p className="text-blue-300 font-mono text-[10px]">
+                            Current: {currentTimeString}
+                        </p>
+                    )}
+                    {preloadedCount > 0 && (
+                        <p className="text-purple-400 text-[10px]">
+                            📦 Preloaded: {preloadedCount} {isTemporalLayer ? "frames" : "days"}
+                        </p>
+                    )}
+                </div>
+
+                <div className="text-xs text-gray-500 border-t border-gray-700 pt-2">
+                    💡 Scrub timeline - images preload in background for smooth playback
+                </div>
+                <div className="text-xs text-gray-500 border-t border-gray-700 pt-2">
+                    ⚙️ Resolution:
+                    {resolution === "potato" ? "Potato (Fastest)" : resolution === "low" ? " Low (Fast)" : resolution === "medium" ? " Medium (recommended)" : " High (Requires high speed internet)"}
+                    <select
+                        value={resolution}
+                        onChange={HandleResolutionChange}
+                        className="w-full px-3 py-2 bg-gray-800 border border-gray-600 rounded text-white focus:border-blue-500 focus:outline-none"
+                        >
+                        <option label="Potato" value="potato" />
+                        <option label="Low" value="low" />
+                        <option label="Medium" value="medium" />
+                        <option label="High" value="high" />
+                    </select>
+                </div>
+            </div>
+        </div>
+    );
+}
+*/
