+++ Stellar-View/src/index.css
@@ -1,8 +1,4 @@
 @import "tailwindcss";
-html, body {
-    width: 100%;
-    overflow-x: hidden;
-}
 @layer components {
     .button-component {
         position: relative;
